Question conversion to convert module
